pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "labyrinth"
version = "0.1.0"
authors = [
    {name = "Dario Palladino", email = "dario@palladinomail.com"},
]
description = "High-level Python API wrapper for Google's A2A SDK - simplifying agent-to-agent communication"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Communications",
    "Topic :: Internet",
    "Topic :: System :: Distributed Computing",
]
keywords = ["a2a", "agent", "communication", "ai", "google", "sdk"]
dependencies = [
    "a2a-sdk>=0.3.0",
    "pydantic>=2.0.0",
    "httpx>=0.24.0",
    "python-dotenv>=0.19.0",
    "structlog>=22.0.0",
    "typing-extensions>=4.0.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "aiohttp>=3.8.0",
    "azure-identity>=1.15.0",
    "PyJWT>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.10.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0",
    "pre-commit>=2.20.0",
]
docs = [
    "sphinx>=5.0.0",
    "sphinx-rtd-theme>=1.0.0",
    "myst-parser>=0.18.0",
]

[project.urls]
"Homepage" = "https://github.com/dariopalladino/labyrinth"
"Bug Reports" = "https://github.com/dariopalladino/labyrinth/issues"
"Source" = "https://github.com/dariopalladino/labyrinth"
"Documentation" = "https://github.com/dariopalladino/labyrinth/README.md"

[project.scripts]
labyrinth-registry = "labyrinth.cli.registry:registry"

[tool.setuptools.packages.find]
where = ["."]
include = ["labyrinth*"]
exclude = ["tests*", "docs*", "examples*"]


[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers"
testpaths = [
    "tests",
]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "a2a.*",
]
ignore_missing_imports = true
